+++ WLEditor/Toolbox/ToolboxForm.cs
@@ -95,7 +95,7 @@ namespace WLEditor
 
 		void ToolBoxTile16x16MouseMove(object sender, TileEventArgs e)
 		{
-			int tileIndex = e.TileX + e.TileY * 8;
+			byte tileIndex = (byte)(e.TileX + e.TileY * 8);
 			var tileInfo = Level.GetTileInfo(tileIndex, tiles16x16PictureBox.SwitchType);
 			toolStripStatusLabel1.Text = string.Format($"{tileIndex:X2} {tileInfo.Text}");
 			statusStrip1.Visible = true;
